refactor: deduplicate button bar and sanitization patterns
Closes #415
