+++ src/utils/terminal-subsystem.js
@@ -41,7 +41,7 @@ export {
 } from './terminal-events.js';
 
 // ── dom-facades (board-view) ────────────────────────────────────────
-export { _el, renderButtonBar, renderList } from './dom-facades.js';
+export { _el, renderButtonBar, buildDomainButtonBar, renderList } from './dom-facades.js';
 
 // ── terminal-factory (board-view) ───────────────────────────────────
 export {
